Integrate Unit Tests for Create, GetAll, GetInvoiceByNumber, and Delete Methods with Exception Handling

+++ src/main/java/com/cato/msa/invoices/service/impl/InvoiceHeaderServiceImpl.java
@@ -1,8 +1,8 @@
 package com.cato.msa.invoices.service.impl;
 
 import com.cato.msa.invoices.domain.InvoiceHeader;
-import com.cato.msa.invoices.exception.NotContentException;
-import com.cato.msa.invoices.exception.NotFoundException;
+import com.cato.msa.invoices.exceptions.NotContentException;
+import com.cato.msa.invoices.exceptions.NotFoundException;
 import com.cato.msa.invoices.repository.InvoiceHeaderRepository;
 import com.cato.msa.invoices.service.InvoiceHeaderService;
 import org.springframework.stereotype.Service;
@@ -40,6 +40,10 @@ public class InvoiceHeaderServiceImpl implements InvoiceHeaderService {
 
     @Override
     public void deleteById(Long id) {
+        boolean exists = invoiceHeaderRepository.existsById(id);
+        if(!exists){
+            throw new NotFoundException("Invoice with id '" + id + "' was not found");  
+        }
         invoiceHeaderRepository.deleteById(id);
     }
